lsp/cache: Override the registry Fetch method within LSP
LSP uses modpkgload to load packages. modpkgload uses modcache to locate
modules within the CUE_CACHE_DIR. When modpkgload parses .cue files, it
uses the ImportsOnly parse option.

When the LSP loads some package P1 in module M1, and P1 imports some
other package P2 in module M2 and that module exists in the
CUE_CACHE_DIR, the LSP needs to module both modules and packages, and
have access to the full ASTs for all files within those packages. This
is required for the jump-to-dfn analysis, for example.

But, calling modpkgload to load P1 in the main module M1 would find the
import of P2 and follow it, locating M2 via modcache, but would only
parse the imports of P2 itself. Previously, we were detecting this
situation, and having the LSP reload P2 in M2 fully in order to ensure
we parse its cue files fully.

However, if we intercept the results of modcache.Fetch and switch out
the FS, we can ensure that when modpkgload subsequently reads the CUE
files from that FS, we can manipulate the parsing mode, and cache the
AST if possible. This ensures that the parsing, and caching of the AST,
is identical whether the package is being loaded within the main module,
or some external module. Consequently, the LSP itself need not worry
about detecting new packages that are also external.

Signed-off-by: Matthew Sackman <matthew@cue.works>
Change-Id: Ia0ce8055d9ac41f3264c3ba8cfdcbc6725612dcc
Dispatch-Trailer: {"type":"trybot","CL":1220490,"patchset":2,"ref":"refs/changes/90/1220490/2","targetBranch":"master"}