doc/spec: use "custom errors" instead of "user errors"
The spec refers to "user-generated errors" and "user errors". This is
highly confusing and ambiguous; a user error (to most people) is an
error in the way in which something is used, rather than the
specification of the error message that should be show when something
is, well, used the wrong way!

Instead prefer the term "custom error" which speaks much more clearly to
the goal of the user.

The updates in cue/testdata/builtins/error.txtar are just for
consistency in terms of naming.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I9f76077d9076ba1b1a8c1da1bf701416c9e96abf
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/cue/+/1222841
Reviewed-by: Marcel van Lohuizen <mpvl@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>