encoding/jsonschema: use ast.Expr rather than map
The current `item.generate` method returns a `map[string]any`
which is quite limiting:
- we can't choose the order of fields
- we can't return `true` or `false` when that might be appropriate.
Instead, change the interface so it returns an `ast.Expr`
which gives it more freedom.

While we're about it, we can also use an `ast.Expr` instead
of `json.RawMessage` in `itemConst` and `itemEnum`, which
means we can preserve the original formatting of numeric constants.

This means we can and do now sort the fields according to order
preference.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ibaa534f9906d7b6f1d7628b4aa824b394d3c9ad7
Dispatch-Trailer: {"type":"trybot","CL":1224199,"patchset":6,"ref":"refs/changes/99/1224199/6","targetBranch":"master"}