internal/core/adt: fix disjunction related mem mgmt bug
When a node with a disjunction had a cycle with itself,
the disjunction could be freed before
`finalizeDisjunctions` in the outer node was called.
This caused a nil pointer panic.

We now don't process disjunctions if the enclosing
node has a positive refCount.

Issue #4119

Signed-off-by: Marcel van Lohuizen <mpvl@gmail.com>
Change-Id: I2f99520cda61798b82e28838b21c8d9b295490f8
Dispatch-Trailer: {"type":"trybot","CL":1224531,"patchset":1,"ref":"refs/changes/31/1224531/1","targetBranch":"master"}