internal/core/compile: don't panic on an alias cycle with a dynamic field
Both #2094 and #2527 were alternative versions of one underlying bug,
where using a dynamic field whose value is an alias to itself
would reach an "unreachable" panic. This happens with both the old
and default alias syntax, as well as aliasv2, so test both forms.

Note that we add the tests in the same commit as the fix,
as otherwise the entire test file causes a panic.
I did verify, one by one, that each of the two reproducers panics
with both the old and new syntax without the fix.

Fixes #2094.
Fixes #2527.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ie7dd36826ce6c32924936b9f492af69c62575840
Dispatch-Trailer: {"type":"trybot","CL":1226414,"patchset":2,"ref":"refs/changes/14/1226414/2","targetBranch":"master"}